Refactor service layer to use DTOs (#90)

+++ src/main/java/com/lukaskucera/numberneighbors/service/PlayerServiceImpl.java
@@ -1,19 +1,24 @@
 package com.lukaskucera.numberneighbors.service;
 
+import com.lukaskucera.numberneighbors.dto.AuthDTO;
+import com.lukaskucera.numberneighbors.dto.PlayerDTO;
+import com.lukaskucera.numberneighbors.dto.UserGameDTO;
 import com.lukaskucera.numberneighbors.entity.GameEntity;
 import com.lukaskucera.numberneighbors.entity.PlayerEntity;
+import com.lukaskucera.numberneighbors.exception.GameNotFoundException;
 import com.lukaskucera.numberneighbors.exception.GamePopulatedException;
-import com.lukaskucera.numberneighbors.exception.PlayerIdMissingInJwtTokenClaimsException;
 import com.lukaskucera.numberneighbors.exception.PlayerNameAlreadyExistsException;
 import com.lukaskucera.numberneighbors.exception.PlayerNotFoundException;
+import com.lukaskucera.numberneighbors.repository.GameRepository;
 import com.lukaskucera.numberneighbors.repository.PlayerRepository;
+import java.util.Optional;
 import java.util.Set;
+import java.util.stream.Collectors;
 import javax.transaction.Transactional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -23,48 +28,67 @@ public class PlayerServiceImpl implements PlayerService {
     PlayerServiceImpl.class
   );
 
+  private final GameServiceImpl gameService;
+
+  private final GameRepository gameRepository;
   private final PlayerRepository playerRepository;
 
-  public PlayerServiceImpl(PlayerRepository playerRepository) {
+  public PlayerServiceImpl(
+    GameServiceImpl gameService,
+    GameRepository gameRepository,
+    PlayerRepository playerRepository
+  ) {
+    this.gameService = gameService;
+    this.gameRepository = gameRepository;
     this.playerRepository = playerRepository;
   }
 
   @Override
-  public Long getPlayerIdFromToken(JwtAuthenticationToken jwtToken) {
-    final Long playerId = (Long) jwtToken
-      .getToken()
-      .getClaims()
-      .get("playerId");
+  public PlayerDTO getPlayerById(AuthDTO auth, Long id) {
+    checkPlayerAccess(auth, id);
 
-    if (playerId == null) {
-      throw new PlayerIdMissingInJwtTokenClaimsException(jwtToken);
-    }
-
-    return playerId;
-  }
-
-  @Override
-  public PlayerEntity getPlayerById(Long id) {
-    return playerRepository
-      .findById(id)
+    return getPlayerEntity(id)
+      .map(PlayerDTO::fromPlayer)
       .orElseThrow(() -> new PlayerNotFoundException(id));
   }
 
   @Override
-  public PlayerEntity getPlayerByJwtToken(JwtAuthenticationToken jwtToken) {
-    final Long playerId = getPlayerIdFromToken(jwtToken);
+  public Set<PlayerDTO> getPlayersByGameId(AuthDTO auth, Long gameId) {
+    gameService.checkGameAccess(auth, gameId);
 
-    return getPlayerById(playerId);
+    return playerRepository
+      .findPlayerEntitiesByGameId(gameId)
+      .stream()
+      .map(PlayerDTO::fromPlayer)
+      .collect(Collectors.toSet());
   }
 
   @Override
-  public Set<PlayerEntity> getPlayersByGameId(Long gameId) {
-    return playerRepository.findPlayerEntitiesByGameId(gameId);
+  public UserGameDTO getPlayerUserGameByGameId(AuthDTO auth, Long gameId) {
+    gameService.checkGameAccess(auth, gameId);
+
+    return getPlayerEntity(auth.playerId())
+      .map(UserGameDTO::fromPlayer)
+      .orElseThrow(() -> new PlayerNotFoundException(auth.playerId()));
   }
 
   @Override
   @Transactional
-  public PlayerEntity newPlayer(String name, GameEntity game) {
+  public PlayerDTO newPlayer(Long gameId, String name) {
+    final GameEntity game = gameRepository
+      .findById(gameId)
+      .orElseThrow(() -> new GameNotFoundException(gameId));
+
+    checkGamePlayersState(name, game);
+
+    final PlayerEntity player = createPlayer(name, game);
+
+    gameService.sendPayloadToPlayers(player.getGame());
+
+    return PlayerDTO.fromPlayer(player);
+  }
+
+  public void checkGamePlayersState(String name, GameEntity game) {
     final Set<PlayerEntity> players = game.getPlayers();
 
     if (players.size() >= GameServiceImpl.PLAYER_COUNT) {
@@ -73,6 +97,7 @@ public class PlayerServiceImpl implements PlayerService {
         name,
         game.getId()
       );
+
       throw new GamePopulatedException(game.getId());
     }
 
@@ -85,9 +110,12 @@ public class PlayerServiceImpl implements PlayerService {
         name,
         game.getId()
       );
+
       throw new PlayerNameAlreadyExistsException(game.getId(), name);
     }
+  }
 
+  public PlayerEntity createPlayer(String name, GameEntity game) {
     final PlayerEntity player = new PlayerEntity(name, game);
     game.addPlayer(player);
 
@@ -97,7 +125,9 @@ public class PlayerServiceImpl implements PlayerService {
   }
 
   @Override
-  public void deletePlayerById(Long id) {
+  public void deletePlayerById(AuthDTO auth, Long id) {
+    checkPlayerAccess(auth, id);
+
     try {
       playerRepository.deleteById(id);
     } catch (EmptyResultDataAccessException e) {
@@ -106,19 +136,19 @@ public class PlayerServiceImpl implements PlayerService {
   }
 
   @Override
-  public void checkPlayerAccess(
-    Long playerId,
-    JwtAuthenticationToken jwtToken
-  ) {
-    final Long claimedPlayerId = getPlayerIdFromToken(jwtToken);
-
-    if (claimedPlayerId == null || !claimedPlayerId.equals(playerId)) {
+  public void checkPlayerAccess(AuthDTO auth, Long playerId) {
+    if (!auth.playerId().equals(playerId)) {
       logger.info(
         "Authenticated to access the player {} not {}",
-        claimedPlayerId,
+        auth.playerId(),
         playerId
       );
+
       throw new AccessDeniedException("Access denied to player " + playerId);
     }
+  }
+
+  private Optional<PlayerEntity> getPlayerEntity(Long id) {
+    return playerRepository.findById(id);
   }
 }
